[frontend] WIP: Homepage hero section text
Used Vue and formatted it with CSS to add "Welcome to Chigame"
to the hero section of the homepage. The Vue was really unnecessary
so I think I may remove it later on. It was a good learning
experience to try to make an App in Vue though. Hero section still
needs the images added and then we can fix spacing as well.

WIP: #518