[knowledge-base] Create Fixture with Dummy Data (#756)
* [knowledge-base] create fixture

- create user/game objects
- create guide objects
- create reviewfeedback objects
- modify models.py to make likes/favorites optional

Closes: #752

* [knowledge-base] feedback fixture change

Modify feedback objects such that Guide 1 has both "requested change" and
"approved". Helpful for guide history if we need it oneday.

* [knowledge-base] add fixture readme

Added a fixture readme file that contains the instructions for
loading the fixture, and the detailed info for the example objects.

For: #752

* [knowledge-base] update readme

Remove backslash from the command.

For: #752

* [knowledge-base] modify fixture README.md

Add a line break after the password to avoid confusion.

For: #752

---------

Co-authored-by: Yi Zou <yizou@Yis-MacBook-Air.local>
Co-authored-by: jacqsmei <54288733+jacqsmei@users.noreply.github.com>