[Knowledge Base] Created white-label mode for guides (#751)
* [knowledge-base] Added barebones markdown sandbox

This should be a working first implementation of the sandbox; however, before finalizing
this commit as a PR, I'd like another team member to pull and run this branch. Since
we're adding a new library, the dependencies of this branch could be inconsistent across
machines at this point.

WIP: #513

* [knowledge-base] Designed basic sandbox UI, fixed list rendering

Given that our sandbox renders HTML tags, I wanted the UI to
clearly delineate what is being rendered vs. what is part
of the sandbox itself.

Completes task 1 of #513

* [knowledge-base] Updated README.md to reflect new changes

* [knowledge-base] Moved sandbox styling to external file

Fixes #519

* [knowledge-base] implement multi-page markdown rendering with section navigation WIP: #513

* [knowledge-base] add white-label mode for minimal UI WIP: #640

* [knowledge-base] revert sandbox demo code after conflict #640

* chore: apply pre-commit fixes and Black formatting

* fixed lint and merge issues

* [knowledge-base] ignore section parameter in minimal mode WIP #640

---------

Co-authored-by: Luca Dovichi <lucadovichi@gmail.com>