[API] Design and Implement POST endpoint to post a MetricScore for LEADERBOARDS (#968)
* [API] WIP: Added leaderboard post endpoint to
post scores to the leaderboard for a game.

Works locally, but needs to be tested after
PR #777 has been merged. Addresses issue #699.

* [api] WIP: Changed retrieving user during leaderboard post

* Removed unused testing libraries in views file

* [api] #968, wip: resolved merge conflicts and solved pre-commit formatting

* [api] wip: continuing progress on issue #699.

I was facing migration issues with the leaderboard
app since I was using obsolete migrations. I had to
delete the old migrations and create new ones.

The POST endpoint for the leaderboard now works.

* [api] wip: fixed linter errors and made the branch
up to date with dev. closes #968

* [api] wip: closes #699. fixed linter tests.

---------

Co-authored-by: theworldofobi <112725097+theworldofobi@users.noreply.github.com>
Co-authored-by: Sam Xie <72187977+MuyanXie@users.noreply.github.com>