[frontend] Homepage reacts to AuthenticationStatus
I added a CheckAuth API to check if the user is authenticated. Then I
wrote a Vue composable (reusable piece of code) that checks if users
are authenticated or not. The code changes the look of the homepage
and navbar based on whether or not the user is authenticated. I also
added a Sign Out button with signout functionality to the navbar in
order to see these changes more easily.

Closes: #1437